(refactor) reorganized claims dir
