+++ packages/esm-billing-app/src/claims/claims-dashboard/claims-breakdown/claims-breakdown.scss
