+++ packages/esm-billing-app/src/claims/claims-dashboard/claims-header/claims-header.component.tsx
@@ -5,8 +5,8 @@ import { useTranslation } from 'react-i18next';
 import { ExtensionSlot, usePatient, showModal } from '@openmrs/esm-framework';
 import { ErrorState } from '@openmrs/esm-patient-common-lib';
 import styles from './claims-header.scss';
-import { useBill } from '../../../../billing.resource';
-import { LineItem } from '../../../../types';
+import { useBill } from '../../../billing.resource';
+import { LineItem } from '../../../types';
 import ClaimsTable from '../claims-table.component.tsx/claims-table.component';
 import { ClaimsBreakDown } from '../claims-breakdown/claims-breakdown.component';
 import ClaimsForm from '../claims-form/claims-form.component';
