+++ packages/esm-billing-app/src/claims/claims-dashboard/claims-header/claims-header.scss
