+++ packages/esm-billing-app/src/claims/claims-dashboard/claims-table.component.tsx/claims-table.scss
