+++ packages/esm-billing-app/src/root.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { BillingDashboard } from './billing-dashboard/billing-dashboard.component';
 import Invoice from './invoice/invoice.component';
-import ClaimScreen from './invoice/claims/claims-dashboard/claims-dashboard.component';
+import ClaimScreen from './claims/claims-dashboard/claims-dashboard.component';
 
 const RootComponent: React.FC = () => {
   const baseName = window.getOpenmrsSpaBase() + 'home/billing';
